Fix a monster_type misbuild with clang-3.3.
Again, a magic value of -1.  It was also wrongly checked in a few error
handling routines, rendering them inoperative even with gcc.
(cherry picked from commit 5f493bf4362078a61db8c6a62e16e1faa9a586fe)

+++ crawl-ref/source/dungeon.cc
@@ -4418,7 +4418,7 @@ monster* dgn_place_monster(mons_spec &mspec, coord_def where,
                            bool force_pos, bool generate_awake, bool patrolling)
 {
 #if TAG_MAJOR_VERSION == 34
-    if (mspec.type == -1) // or rebuild the des cache
+    if ((int)mspec.type == -1) // or rebuild the des cache
         return 0;
 #endif
     if (mspec.type == MONS_NO_MONSTER)
